System Admin Code Reference
I've put together a System Administration Commands Reference on GitHub to make life a bit easier for fellow system administrators and beginners. It's quick access to essential Linux commands with straightforward explanations. Feel free to dive in, and use it for your daily tasks, and if you've got more commands to add or tweaks to suggest, let's make this a shared space where we all benefit from each other's Linux wisdom!